+++ src/panels/lovelace/header-footer/hui-buttons-header-footer.ts
@@ -1,97 +1,16 @@
-import {
-  html,
-  LitElement,
-  TemplateResult,
-  customElement,
-  css,
-  CSSResult,
-  queryAll,
-} from "lit-element";
-import "@material/mwc-ripple";
+import { customElement } from "lit-element";
 
-import "../../../components/entity/state-badge";
-import "../../../components/ha-card";
-import "../../../components/ha-icon";
-import "../components/hui-warning-element";
-
-import { HomeAssistant } from "../../../types";
+import { HuiButtonsBase } from "../components/hui-buttons-base";
 import { LovelaceHeaderFooter } from "../types";
 import { ButtonsHeaderFooterConfig } from "./types";
-import { EntityConfig } from "../entity-rows/types";
 import { processConfigEntities } from "../common/process-config-entities";
-import { toggleEntity } from "../common/entity/toggle-entity";
-import { computeTooltip } from "../common/compute-tooltip";
-// tslint:disable-next-line: no-duplicate-imports
-import { StateBadge } from "../../../components/entity/state-badge";
 
 @customElement("hui-buttons-header-footer")
-export class HuiButtonsHeaderFooter extends LitElement
+export class HuiButtonsHeaderFooter extends HuiButtonsBase
   implements LovelaceHeaderFooter {
-  public static getStubConfig(): object {
-    return { entities: [] };
-  }
-
-  private _configEntities?: EntityConfig[];
-  private _hass?: HomeAssistant;
-  @queryAll("state-badge") private _badges!: StateBadge[];
-
   public setConfig(config: ButtonsHeaderFooterConfig): void {
     this._configEntities = processConfigEntities(config.entities);
     this.requestUpdate();
-  }
-
-  set hass(hass: HomeAssistant) {
-    this._hass = hass;
-    this._badges.forEach((badge, index: number) => {
-      badge.hass = hass;
-      badge.stateObj = hass.states[this._configEntities![index].entity];
-    });
-  }
-
-  protected render(): TemplateResult | void {
-    return html`
-      ${(this._configEntities || []).map((entityConf) => {
-        const stateObj = this._hass!.states[entityConf.entity];
-        if (!stateObj) {
-          return html`<div class='missing'><iron-icon icon="hass:alert"></div>`;
-        }
-
-        return html`
-          <div>
-            <state-badge
-              title=${computeTooltip(this._hass!, entityConf)}
-              @click=${this._toggle}
-              .hass=${this._hass}
-              .stateObj=${stateObj}
-              .overrideIcon=${entityConf.icon}
-              .overrideImage=${entityConf.image}
-              stateColor
-              tabindex="0"
-            ></state-badge>
-            <mwc-ripple unbounded></mwc-ripple>
-          </div>
-        `;
-      })}
-    `;
-  }
-
-  private async _toggle(ev) {
-    await toggleEntity(this._hass!, ev.target.stateObj.entity_id);
-  }
-
-  static get styles(): CSSResult {
-    return css`
-      :host {
-        display: flex;
-        justify-content: space-evenly;
-      }
-      .missing {
-        color: #fce588;
-      }
-      state-badge {
-        cursor: pointer;
-      }
-    `;
   }
 }
